# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "post": {
        "operationId": "UsersController_add",
        "summary": "Add a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_getAll",
        "summary": "Get all users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateRoot",
        "summary": "Update existing user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/add": {
      "post": {
        "operationId": "UsersController_add",
        "summary": "Add a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/authenticate": {
      "post": {
        "operationId": "UsersController_authenticate",
        "summary": "Authenticate user and return JWT token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthTokenResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/users/update": {
      "post": {
        "operationId": "UsersController_update",
        "summary": "Update existing user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/details/{id}": {
      "get": {
        "operationId": "UsersController_details",
        "summary": "Get user details by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_details",
        "summary": "Get user details by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/get_self": {
      "get": {
        "operationId": "UsersController_getSelf",
        "summary": "Get logged in user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current authenticated user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "LMS Backend API",
    "description": "NestJS REST API documentation for LMS Backend Management System",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "AddUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "Email address of the user"
          },
          "password": {
            "type": "string",
            "example": "Pass@1234",
            "description": "User password (at least 8 characters, one alphabet, one number, and one special character)"
          },
          "role": {
            "type": "string",
            "example": "student",
            "enum": [
              "student",
              "faculty",
              "admin"
            ],
            "description": "User role"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "Unique user identifier"
          },
          "roll_no": {
            "type": "string",
            "example": "STU001",
            "description": "User roll number"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "User full name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "User email"
          }
        },
        "required": [
          "id",
          "roll_no",
          "name",
          "email"
        ]
      },
      "AuthenticateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "User email or roll number"
          },
          "password": {
            "type": "string",
            "example": "Pass@1234",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserAuthTokenResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT Authentication Bearer token"
          },
          "data": {
            "description": "Authenticated user info",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          }
        },
        "required": [
          "token",
          "data"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "example": 1,
            "description": "User ID to update"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Updated full name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "Updated email address"
          },
          "password": {
            "type": "string",
            "example": "Pass@1234",
            "description": "Updated password"
          },
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "10-digit mobile phone number"
          }
        },
        "required": [
          "id"
        ]
      }
    }
  }
}